fixed fs

+++ include/gfc_fs.h
@@ -55,7 +55,7 @@ void
 gfc_fs_iterate(const char* path, user_data data, void (*resolve)(const char*, user_data));
 
 void
-gfc_fs_write(const char* path, const byte* content, size_t len);
+gfc_fs_write(const char* path, const byte* content, ulong len);
 
 #ifdef __cplusplus
 }
